// dfp-piao-fe/src/app/shared/models/classes/fase-dto.ts
import { CampiTecniciDTO } from './campi-tecnici-dto';
import { AttoreDTO } from './attore-dto';
import { AttivitaDTO } from './attivita-dto';
import { BaseMongoAttivitaDTO } from './base-mongo-attivita-dto';

export class FaseDTO extends CampiTecniciDTO {
  id?: number;
  idSezione22?: number;
  denominazione?: string;
  descrizione?: string;
  tempi?: string;
  attore?: AttoreDTO;
  attivita?: BaseMongoAttivitaDTO;
}